+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # `versions` Changelog
 
-## Unreleased
+## 2.0.0 (2020-10-21)
 
 #### Changed